scrapper/requirements.txt
requests
lxml
sqlalchemy
awesome-slugify=1.6.5